+++ src/main/java/ru/vyarus/dropwizard/orient/OrientServerBundle.java
@@ -2,10 +2,11 @@ package ru.vyarus.dropwizard.orient;
 
 import io.dropwizard.Configuration;
 import io.dropwizard.ConfiguredBundle;
+import io.dropwizard.jersey.validation.Validators;
 import io.dropwizard.jetty.NonblockingServletHolder;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
-import org.hibernate.validator.internal.engine.resolver.DefaultTraversableResolver;
+import org.hibernate.validator.internal.engine.resolver.JPATraversableResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.vyarus.dropwizard.orient.configuration.HasOrientServerConfiguration;
@@ -15,11 +16,11 @@ import ru.vyarus.dropwizard.orient.configuration.deserializer.NetworkProtocolDes
 import ru.vyarus.dropwizard.orient.configuration.deserializer.ParameterDeserializer;
 import ru.vyarus.dropwizard.orient.health.OrientServerHealthCheck;
 import ru.vyarus.dropwizard.orient.internal.EmbeddedOrientServer;
+import ru.vyarus.dropwizard.orient.internal.TraverseAllResolver;
 import ru.vyarus.dropwizard.orient.support.ConsoleCommand;
 import ru.vyarus.dropwizard.orient.support.OrientServlet;
 
 import javax.validation.TraversableResolver;
-import java.lang.reflect.Field;
 
 /**
  * Bundle starts embedded orient server. Application configuration object must implement
@@ -92,29 +93,28 @@ public class OrientServerBundle<T extends Configuration & HasOrientServerConfigu
      * Orientdb-object module includes hiberate-jpa-api into classpath and hibernate-validator
      * starts to think that JPA is available (if jpa available then validated objects must be checked with
      * traversable provider to avoid lazy init exceptions).
-     * But orient's OJPAPersistenceProvider.getProviderUtil() simply throws exception.
-     * So the only way to resolve problem is to disabled incorrectly detected jpa support in
-     * {@link org.hibernate.validator.internal.engine.resolver.DefaultTraversableResolver}.
-     * It's not a hack: in normal case, validator also use dummy impl and only if
-     * javax.persistence.Persistence class found in classpath use complete impl.. so we just correct
-     * behaviour here.
+     * But orient's {@code OJPAPersistenceProvider.getProviderUtil()} simply throws exception.
+     * <p>
+     * The only way to resolve problem is to manually register correct {@link TraversableResolver}.
+     * (see {@link org.hibernate.validator.internal.engine.resolver.TraversableResolvers#getDefault()} for
+     * default resolver selection logic).
+     * <p>
+     * It's not a hack: in normal case, validator use the same implementation, so it's only behaviour fix due to
+     * additional jpa jar appeared in classpath (validator is wrong assuming it means jpa will be really used).
+     * This substitution could only override manually registered factory and to avoid this create this
+     * factory with correct resolver.
      * <p>
      * Note that it can't cause side effects because hibernate is actually not used.
      */
     private void recoverValidatorBehaviour(final Bootstrap<?> bootstrap) {
         final TraversableResolver resolver = bootstrap.getValidatorFactory().getTraversableResolver();
-        if (isObjectOrientUsed() && resolver instanceof DefaultTraversableResolver) {
-            logger.debug("Removing default TraversableResolver jpa support to fix hibernate validator");
-            // only default DefaultTraversableResolver implicitly enables JPA factory
-            try {
-                final Field field = resolver.getClass().getDeclaredField("jpaTraversableResolver");
-                field.setAccessible(true);
-                // disable jpa support (incorrectly detected)
-                field.set(resolver, null);
-                field.setAccessible(false);
-            } catch (Exception e) {
-                throw new IllegalStateException("Failed disable JPA support for traversable resolver", e);
-            }
+        if (isObjectOrientUsed() && resolver instanceof JPATraversableResolver) {
+            logger.debug("Overriding incorrectly configured ValidatorFactory to remove jpa support");
+            // repalce JPA traversible resolver with default implementation (it would be uset automatically
+            // if jpa jar would not appear in classpath
+            bootstrap.setValidatorFactory(Validators.newConfiguration()
+                    .traversableResolver(new TraverseAllResolver())
+                    .buildValidatorFactory());
         }
     }