+++ src/test/groovy/ru/vyarus/dropwizard/orient/LuceneTest.groovy
@@ -45,7 +45,7 @@ class LuceneTest extends AbstractTest {
         when: "accessing health checks"
         data = new URL("http://localhost:8081/healthcheck").getText()
         then: "all good"
-        data && data.contains('"orient-server":{"healthy":true,"message":"OK"}')
+        data && data.contains('"orient-server":{"healthy":true,"message":"OK"')
 
         when: "accessing orient servlet"
         data = new URL("http://localhost:8081/orient").getText()